requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
jupyter>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
black>=23.0.0
azure-storage-blob>=12.16.0
azure-core>=1.24.0
selenium>=4.14.0
webdriver-manager>=4.0.1
agent-framework-azure-ai
azure-identity
openai
azure-ai-projects>=2.0.0b1